+++ src/lib/ares_query.c
@@ -74,7 +74,7 @@ ares_status_t ares_query_qid(ares_channel channel, const char *name,
   /* Compose the query. */
   rd = !(channel->flags & ARES_FLAG_NORECURSE);
   status = (ares_status_t)ares_create_query(name, dnsclass, type, id, rd, &qbuf,
-              &qlen, (channel->flags & ARES_FLAG_EDNS) ? channel->ednspsz : 0);
+              &qlen, (channel->flags & ARES_FLAG_EDNS) ? (int)channel->ednspsz : 0);
   if (status != ARES_SUCCESS)
     {
       if (qbuf != NULL) ares_free(qbuf);
